requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Database and ORM
sqlalchemy==2.0.23
sqlite3

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# LLM APIs
google-generativeai==0.3.2
anthropic==0.7.8
openai==1.3.6

# ML and embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3

# Web scraping
aiohttp==3.9.1
beautifulsoup4==4.12.2
requests==2.31.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Email
aiosmtplib==3.0.1

# Payments
stripe==7.8.0

# Utilities
pydantic==2.5.0
asyncio
logging
datetime
pathlib
re
json
pickle
hashlib
unicodedata
enum
dataclasses
typing